add api cate

+++ src/main/java/com/trungtamjava/CuDau/ApiController/CategoryAPIController.java
@@ -3,7 +3,11 @@ package com.trungtamjava.CuDau.ApiController;
 import java.util.List;
 import java.util.function.LongFunction;
 
+import javax.xml.ws.Response;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,6 +18,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.trungtamjava.CuDau.Dto.CategoryDto;
+import com.trungtamjava.CuDau.Entity.CategoryEntity;
+import com.trungtamjava.CuDau.Repository.CategoryRepository;
 import com.trungtamjava.CuDau.Service.CategoryService;
 
 @RestController
@@ -23,32 +29,47 @@ public class CategoryAPIController {
 	@Autowired 
 	CategoryService categoryService;
 	
+	@Autowired
+	CategoryRepository categoryRepository;
+	
 	@PostMapping("/admin/category/add")
-	public CategoryDto add(@RequestBody CategoryDto categoryDto) {
-		categoryService.add(categoryDto);
-		return categoryDto;
+	public ResponseEntity<CategoryDto> add(@RequestBody CategoryDto categoryDto) {
+		return new ResponseEntity<CategoryDto>(categoryService.add(categoryDto), HttpStatus.OK);
 	}
 	
 	@PutMapping("/admin/category/update")
-	public CategoryDto update(@RequestBody CategoryDto categoryDto) {
-		categoryService.update(categoryDto);
-		return categoryDto;
+	public ResponseEntity<CategoryDto> update(@RequestBody CategoryDto categoryForm, @RequestParam(value = "id") Long id) {
+		System.out.println(id);
+		CategoryDto categoryDto= categoryService.getCate(id);
+		if (categoryDto != null) {
+			categoryDto.setName(categoryForm.getName());
+			return new ResponseEntity<CategoryDto>(categoryService.save(categoryDto), HttpStatus.OK);
+		}
+		return new ResponseEntity<CategoryDto>(HttpStatus.NOT_FOUND);
+		
+		
 	}
+
 	
 	@DeleteMapping("/admin/category/delete")
-	public void delete(@RequestParam(value = "id") Long id) {
+	public ResponseEntity<CategoryDto> delete(@RequestParam(value = "id") Long id) {
+		System.out.println(id);
 		CategoryDto categoryDto= categoryService.getCate(id);
-		categoryService.delete(categoryDto);
+		if (categoryDto != null) {
+			categoryService.delete(categoryDto);
+			return new ResponseEntity<CategoryDto>(HttpStatus.OK);
+		}
+		return new ResponseEntity<CategoryDto>(HttpStatus.NOT_FOUND);
 	}
 	
 	@GetMapping("/admin/category/get")
-	public CategoryDto Get(@RequestParam(value = "id") Long id) {
-		return categoryService.getCate(id);
+	public ResponseEntity<CategoryDto> getCate (@RequestParam(value = "id") Long id) {
+		return new ResponseEntity<CategoryDto>(categoryService.getCate(id), HttpStatus.OK);
 	}
 	
-	@GetMapping("/admin/category/search")
-	public List<CategoryDto> search(){
-		return categoryService.getAllCate();
+	@GetMapping("/admin/category/getAll")
+	public ResponseEntity<List<CategoryDto>> search(){
+		return new ResponseEntity<List<CategoryDto>>(categoryService.getAllCate(), HttpStatus.OK);
 	}
 
 }
